完成对HOST的PIO部分的重构
