+++ src/mci_bk/regs.rs
@@ -1,22 +1,14 @@
-use core::{default, ops};
-
 use bitflags::bitflags;
-use crate::mci::{constants::*, err::MCIError};
+use crate::mci_bk::{constants::*, err::{FsdifError, FsdifResult}};
 
 use super::{FlagReg, Reg};
 
-pub type MCIReg = Reg<MCIError>;
-
-impl Clone for MCIReg {
-    fn clone(&self) -> Self {
-        Self::new(self.addr)
-    }
-}
+pub type FsdifReg = Reg<FsdifError>;
 
 // FSDIF_CNTRL_OFFSET x0 Register
 bitflags! {
     #[derive(Clone, Copy)]
-    pub struct MCICtrl: u32 {
+    pub struct FsdifCtrl: u32 {
         const CONTROLLER_RESET = 1 << 0; // RW 复位控制器，除 DMA，FIFO
         const FIFO_RESET = 1 << 1; // RW 复位 FIFO, 1 有效
         const DMA_RESET = 1 << 2; // RW 复位内部 DMA, 1 有效
@@ -35,25 +27,24 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCICtrl {
+impl FlagReg for FsdifCtrl {
     const REG: u32 = FSDIF_CNTRL_OFFSET;
 }
 
 // FSDIF_PWREN_OFFSET 0x4 Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIPwrEn: u32 {
+    pub struct FsdifPwrEn: u32 {
         const ENABLE = 1 << 0; // RW 卡供电开关, 0：关；1：开
     }
 }
 
-impl FlagReg for MCIPwrEn {
+impl FlagReg for FsdifPwrEn {
     const REG: u32 = FSDIF_PWREN_OFFSET;
 }
 
 // FSDIF_CLKDIV_OFFSET 0x8 Register
 bitflags! {
-    pub struct MCIClkDiv: u32 {
+    pub struct FsdifClkDiv: u32 {
         /* CLK_SAMPLE 和 CLK_SAMPLE 必须小于 CLK_DIVIDER */
         const CLK_DIVDER_BIT0 = 1 << 0; /* 时钟分频参数设置，分频参数=2*CLK_DIVIDER */
         const CLK_DIVDER_BIT1 = 1 << 1;
@@ -82,11 +73,11 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIClkDiv {
+impl FlagReg for FsdifClkDiv {
     const REG: u32 = FSDIF_CLKDIV_OFFSET;
 }
 
-impl MCIClkDiv {
+impl FsdifClkDiv {
     pub fn clk_sample_set(x:u32) -> Self {
         Self::from_bits_truncate(set_reg32_bits!(x, 23, 16))
     }
@@ -102,26 +93,25 @@ impl MCIClkDiv {
         Self::clk_divider_set(div)
     }
     pub fn clk_divider_get(div_reg:u32) -> Self {
-        MCIClkDiv::from_bits_truncate(get_reg32_bits!(div_reg, 7, 0))
+        FsdifClkDiv::from_bits_truncate(get_reg32_bits!(div_reg, 7, 0))
     }
 }
 
 // FSDIF_CLKENA_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIClkEn: u32 {
+    pub struct FsdifClkEn: u32 {
         const CCLK_ENABLE = 1 << 0; /* RW 0：Clock disabled；1：Clock enabled */
         const CLKENA_CCLK_LOW_POWER = 1<<16; /* RW 0x0：非低功耗；0x1：低功耗 */
     }
 }
 
-impl FlagReg for MCIClkEn {
+impl FlagReg for FsdifClkEn {
     const REG: u32 = FSDIF_CLKENA_OFFSET;
 }
 
 // FSDIF_TMOUT_OFFSET Register
 bitflags! {
-    pub struct MCITimeout: u32 {
+    pub struct FsdifTimeout: u32 {
         const MAX_DATA_TIMEOUT = 0xffffff; /* RW 读卡超时（以卡时钟为单位） */
         const MAX_RESP_TIMEOUT = 0xff; /* RW 响应超时（以卡时钟为单位） */
         const RESP_TIMEOUT_BIT0 = 1 << 0; /* RW 响应超时的第0位 */
@@ -159,13 +149,13 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCITimeout {
+impl FlagReg for FsdifTimeout {
     const REG: u32 = FSDIF_TMOUT_OFFSET;
 }
 
-impl MCITimeout {
-    pub fn timeout_data(data_timeout:MCITimeout,resp_timeout:MCITimeout) -> MCITimeout{
-        MCITimeout::from_bits_truncate(
+impl FsdifTimeout {
+    pub fn timeout_data(data_timeout:FsdifTimeout,resp_timeout:FsdifTimeout) -> FsdifTimeout{
+        FsdifTimeout::from_bits_truncate(
             (genmask!(31,8) & (data_timeout.bits() << 8)) | 
             (genmask!(7,0) & resp_timeout.bits())
         )
@@ -174,32 +164,20 @@ impl MCITimeout {
 
 // FSDIF_CTYPE_OFFSET Register
 bitflags! {
-    pub struct MCICType: u32 {
+    pub struct FsdifCType: u32 {
         const CARD0_WIDTH1_8BIT = 1 << 16; /* 1: 8-bit mode */
         const CARD0_WIDTH2_4BIT = 1 << 0; /* 1: 4-bit mode */
         const CARD0_WIDTH2_1BIT = 0; /* 0: 1-bit mode */
     }
 }
 
-impl From<u32> for MCICType {
-    fn from(val: u32) -> Self {
-        match val {
-            1 => Self::CARD0_WIDTH2_1BIT,
-            4 => Self::CARD0_WIDTH2_4BIT,
-            8 => Self::CARD0_WIDTH1_8BIT,
-            _ => panic!("Invalid card width: {}", val),
-        }
-    }
-}
-
-impl FlagReg for MCICType {
+impl FlagReg for FsdifCType {
     const REG: u32 = FSDIF_CTYPE_OFFSET;
 }
 
 // FSDIF_INT_MASK_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIIntMask: u32 {
+    pub struct FsdifInt: u32 {
         const CD_BIT = 1 << 0;       /* RW Card detect (CD) */
         const RE_BIT = 1 << 1;       /* RW Response error (RE) */
         const CMD_BIT = 1 << 2;      /* RW Command done (CD) */
@@ -223,13 +201,13 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIIntMask {
+impl FlagReg for FsdifInt {
     const REG: u32 = FSDIF_INT_MASK_OFFSET; // 假设FSDIF_INT_OFFSET是对应的寄存器偏移量
 }
 
 // FSDIF_MASKED_INTS_OFFSET Register
 bitflags! {
-    pub struct MCIMaskedInts: u32 {
+    pub struct FsdifMaskedInts: u32 {
         const CD_BIT = 1 << 0;       /* RW Card detect (CD) */
         const RE_BIT = 1 << 1;       /* RW Response error (RE) */
         const CMD_BIT = 1 << 2;      /* RW Command done (CD) */
@@ -253,13 +231,13 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIMaskedInts {
+impl FlagReg for FsdifMaskedInts {
     const REG: u32 = FSDIF_MASKED_INTS_OFFSET;
 }
 
 // FSDIF_RAW_INTS_OFFSET Register
 bitflags! {
-    pub struct MCIRawInts: u32 {
+    pub struct FsdifRawInts: u32 {
         const CD_BIT = 1 << 0;       /* RW Card detect (CD) */
         const RE_BIT = 1 << 1;       /* RW Response error (RE) */
         const CMD_BIT = 1 << 2;      /* RW Command done (CD) */
@@ -283,14 +261,14 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIRawInts {
+impl FlagReg for FsdifRawInts {
     const REG: u32 = FSDIF_RAW_INTS_OFFSET;
 }
 
 // FSDIF_CMD_OFFSET Register
 bitflags! {
     #[derive(Clone, Copy)]
-    pub struct MCICmd: u32 {
+    pub struct FsdifCmd: u32 {
         const START = 1 << 31;                /* 启动命令 */
         const USE_HOLD_REG = 1 << 29;         /* 0: 旁路HOLD寄存器，1: 使能HOLD寄存器 */
         const VOLT_SWITCH = 1 << 28;          /* 0: 无电压转换，1: 有电压转换 */
@@ -317,41 +295,47 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCICmd {
+impl FlagReg for FsdifCmd {
     const REG: u32 = FSDIF_CMD_OFFSET; // 假设FSDIF_CMD_OFFSET是对应的寄存器偏移量
 }
 
-impl MCICmd {
+impl From<u32> for FsdifCmd {
+    fn from(val: u32) -> Self {
+        FsdifCmd::from_bits_truncate(val)
+    }
+}
+
+impl FsdifCmd {
     pub fn index_set(x:u32) -> Self {
         Self::from_bits_truncate(set_reg32_bits!(x, 5, 0))
     }
     pub fn index_get(&self) -> u32 {
-        self.bits() & genmask!(5, 0)
+        (self.bits() & genmask!(5, 0)) >> 0
     }
 }
 
 /* 1: 流数据传输指令 */
-pub fn cmd_transf_mode_set(reg: MCIReg, mode: u32) {
-    reg.modify_reg::<MCICmd>(|reg| {
-        reg | MCICmd::from_bits_truncate(set_reg32_bits!(mode, 12, 11))
+pub fn cmd_transf_mode_set(reg: FsdifReg, mode: u32) {
+    reg.modify_reg::<FsdifCmd>(|reg| {
+        reg | FsdifCmd::from_bits_truncate(set_reg32_bits!(mode, 12, 11))
     });
 }
 
 /* 命令索引号 */
-pub fn cmd_indx_set(reg: MCIReg, ind: u32) {
-    reg.modify_reg::<MCICmd>(|reg| {
-        reg | MCICmd::from_bits_truncate(set_reg32_bits!(ind, 5, 0))
+pub fn cmd_indx_set(reg: FsdifReg, ind: u32) {
+    reg.modify_reg::<FsdifCmd>(|reg| {
+        reg | FsdifCmd::from_bits_truncate(set_reg32_bits!(ind, 5, 0))
     });
 }
 
-pub fn cmd_indx_get(reg: MCIReg) -> u32 {
-    (reg.read_reg::<MCICmd>() & MCICmd::from_bits_truncate(genmask!(5, 0))).bits()
+pub fn cmd_indx_get(reg: FsdifReg) -> u32 {
+    (reg.read_reg::<FsdifCmd>() & FsdifCmd::from_bits_truncate(genmask!(5, 0))).bits()
 }
 
 // FSDIF_STATUS_OFFSET Register
 bitflags! {
     #[derive(Clone, Copy)]
-    pub struct MCIStatus: u32 {
+    pub struct FsdifStatus: u32 {
         const FIFO_RX = 1 << 0;     /* RO, 达到 FIFO_RX 标记 */
         const FIFO_TX = 1 << 1;     /* RO, 达到 FIFO_TX 标记 */
         const FIFO_EMPTY = 1 << 2;  /* RO, FIFO empty */
@@ -387,25 +371,25 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIStatus {
+impl FlagReg for FsdifStatus {
     const REG: u32 = FSDIF_STATUS_OFFSET;
 }
 
-pub fn cmd_fsm_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIStatus>().bits(), 7, 4)
+pub fn cmd_fsm_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifStatus>().bits(), 7, 4)
 }
 
-pub fn resp_index_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIStatus>().bits(), 16, 11)
+pub fn resp_index_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifStatus>().bits(), 16, 11)
 }
 
-pub fn fifo_cnt_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIStatus>().bits(), 29, 17)
+pub fn fifo_cnt_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifStatus>().bits(), 29, 17)
 }
 
 // FSDIF_FIFOTH_OFFSET Register
 bitflags! {
-    pub struct MCIFifoTh: u32 {
+    pub struct FsdifFifoTh: u32 {
         const DMA_TRANS_MASK = genmask!(30, 28); /* 多次传输的突发大小 */
         const RX_WMARK_MASK = genmask!(27, 16);  /* 当接收数据给卡时FIFO的阈值 */
         const TX_WMARK_MASK = genmask!(11, 0);   /* 当发送数据给卡时FIFO的阈值 */
@@ -439,45 +423,29 @@ bitflags! {
     }
 }
 
-impl MCIFifoTh {
-    pub const RX_WMARK: u32 = 0x7;
-    pub const TX_WMARK: u32 = 0x100;
-    /*
-    trans_size: Burst size of multiple transaction;
-    rx_wmark: FIFO threshold watermark level when receiving data to card.
-    tx_wmark: FIFO threshold watermark level when transmitting data to card
-    */
-    pub fn fifoth(
-        trans_size:MCIFifoThDMATransSize,
-        rx_wmark:u32,
-        tx_wmark:u32
-    ) -> Self {
-        let trans_size:u32 = trans_size.into();
-        (MCIFifoTh::DMA_TRANS_MASK & (trans_size << 28)) |
-        (MCIFifoTh::RX_WMARK_MASK & (rx_wmark << 16)) |
-        (MCIFifoTh::TX_WMARK_MASK & tx_wmark)
-    }
-}
-
-BitsOpsForU32!(MCIFifoTh);
-
-impl FlagReg for MCIFifoTh {
+impl FlagReg for FsdifFifoTh {
     const REG: u32 = FSDIF_FIFOTH_OFFSET;
 }
 
-pub enum MCIFifoThDMATransSize {
-    DMATrans1 = 0b000,
-    DMATrans4 = 0b001,
-    DMATrans8 = 0b010,
-    DMATrans16 = 0b011,
-    DMATrans32 = 0b100,
-    DMATrans64 = 0b101,
-    DMATrans128 = 0b110,
-    DMATrans256 = 0b111
+impl From<u32> for FsdifFifoTh {
+    fn from(val: u32) -> Self {
+        FsdifFifoTh::from_bits_truncate(val)
+    }
 }
 
-impl From<MCIFifoThDMATransSize> for u32 {
-    fn from(val: MCIFifoThDMATransSize) -> Self {
+pub enum FsdifFifoThDmaTransSize {
+    DmaTrans1 = 0b000,
+    DmaTrans4 = 0b001,
+    DmaTrans8 = 0b010,
+    DmaTrans16 = 0b011,
+    DmaTrans32 = 0b100,
+    DmaTrans64 = 0b101,
+    DmaTrans128 = 0b110,
+    DmaTrans256 = 0b111
+}
+
+impl From<FsdifFifoThDmaTransSize> for u32 {
+    fn from(val: FsdifFifoThDmaTransSize) -> Self {
         val as u32
     }
 }
@@ -485,86 +453,83 @@ impl From<MCIFifoThDMATransSize> for u32 {
 pub const FSDIF_RX_WMARK:u32 = 0x7;
 pub const FSDIF_TX_WMARK:u32 = 0x100;
 
-pub fn dma_trans_size_set(reg:MCIReg,size:u32){
-    reg.modify_reg::<MCIFifoTh>(|reg| {
-        reg | MCIFifoTh::from_bits_truncate(set_reg32_bits!(size, 30, 28))
+pub fn dma_trans_size_set(reg:FsdifReg,size:u32){
+    reg.modify_reg::<FsdifFifoTh>(|reg| {
+        reg | FsdifFifoTh::from_bits_truncate(set_reg32_bits!(size, 30, 28))
     });
 }
 
-pub fn rx_mark_size_set(reg:MCIReg,size:u32){
-    reg.modify_reg::<MCIFifoTh>(|reg| {
-        reg | MCIFifoTh::from_bits_truncate(set_reg32_bits!(size, 27, 16))
+pub fn rx_mark_size_set(reg:FsdifReg,size:u32){
+    reg.modify_reg::<FsdifFifoTh>(|reg| {
+        reg | FsdifFifoTh::from_bits_truncate(set_reg32_bits!(size, 27, 16))
     });
 }
 
-pub fn tx_mark_size_set(reg:MCIReg,size:u32){
-    reg.modify_reg::<MCIFifoTh>(|reg| {
-        reg | MCIFifoTh::from_bits_truncate(set_reg32_bits!(size, 11, 0))
+pub fn tx_mark_size_set(reg:FsdifReg,size:u32){
+    reg.modify_reg::<FsdifFifoTh>(|reg| {
+        reg | FsdifFifoTh::from_bits_truncate(set_reg32_bits!(size, 11, 0))
     });
 }
 
-// FSDIF_CARD_DETECT_OFFSET Register
+/// FSDIF_CARD_DETECT_OFFSET Register
 bitflags! {
-    pub struct MCICardDetect: u32 {
+    pub struct FsdifCardDetect: u32 {
         const DETECTED = 1 << 0; /* 1：卡不在位；0：卡在位 */
     }
 }
 
-impl FlagReg for MCICardDetect {
+impl FlagReg for FsdifCardDetect {
     const REG: u32 = FSDIF_CARD_DETECT_OFFSET; // 假设 FSDIF_CARD_DETECT_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_CARD_WRTPRT_OFFSET Register
+/// FSDIF_CARD_WRTPRT_OFFSET Register
 bitflags! {
-    pub struct MCICardWrtp: u32 {
+    pub struct FsdifCardWrtp: u32 {
         const WRITE_PROTECTED = 1 << 0; /* 1：写保护；0：无写保护 */
     }
 }
 
-impl FlagReg for MCICardWrtp {
+impl FlagReg for FsdifCardWrtp {
     const REG: u32 = FSDIF_CARD_WRTPRT_OFFSET; // 假设 FSDIF_CARD_WRTPRT_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_CKSTS_OFFSET Register
+/// FSDIF_CKSTS_OFFSET Register
 bitflags! {
-    pub struct MCIClkSts: u32 {
+    pub struct FsdifClkSts: u32 {
         const READY = 1 << 0; /* CIU 时钟 ready */
     }
 }
 
-impl FlagReg for MCIClkSts {
+impl FlagReg for FsdifClkSts {
     const REG: u32 = FSDIF_CKSTS_OFFSET; // 假设 FSDIF_CKSTS_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_UHS_REG_OFFSET Register
+/// FSDIF_UHS_REG_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIUhsReg: u32 {
+    pub struct FsdifUhsReg: u32 {
         const VOLT_180 = 1 << 0; /* RW 外部调压器接口电压 0: 3.3v, 1: 1.8v */
         const DDR = 1 << 16;     /* RW DDR 模式 */
     }
 }
 
-impl FlagReg for MCIUhsReg {
+impl FlagReg for FsdifUhsReg {
     const REG: u32 = FSDIF_UHS_REG_OFFSET; // 假设 FSDIF_UHS_REG_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_CARD_RESET_OFFSET Register
+/// FSDIF_CARD_RESET_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCICardReset: u32 {
+    pub struct FsdifCardReset: u32 {
         const ENABLE = 1 << 0; /* RW 1：运行；0：复位 */
     }
 }
 
-impl FlagReg for MCICardReset {
+impl FlagReg for FsdifCardReset {
     const REG: u32 = FSDIF_CARD_RESET_OFFSET; // 假设 FSDIF_CARD_RESET_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_BUS_MODE_OFFSET Register
+/// FSDIF_BUS_MODE_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIBusMode: u32 {
+    pub struct FsdifBusMode: u32 {
         const SWR = 1 << 0; /* RW 软复位，复位idma内部寄存器 */
         const FB = 1 << 1;  /* RW 固定burst */
         const DE = 1 << 7;  /* RW idma使能 */
@@ -574,17 +539,17 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIBusMode {
+impl FlagReg for FsdifBusMode {
     const REG: u32 = FSDIF_BUS_MODE_OFFSET; // 假设 FSDIF_BUS_MODE_OFFSET 是对应的寄存器偏移量
 }
 
-pub fn bus_mode_pbl_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIBusMode>().bits(), 10, 8)
+pub fn bus_mode_pbl_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifBusMode>().bits(), 10, 8)
 }
 
-// FSDIF_DMAC_STATUS_OFFSET Register
+/// FSDIF_DMAC_STATUS_OFFSET Register
 bitflags! {
-    pub struct MCIDMACStatus: u32 {
+    pub struct FsdifDmacStatus: u32 {
         const TI = 1 << 0;  /* RW 发送中断。表示链表的数据发送完成 */
         const RI = 1 << 1;  /* RW 接收中断。表示链表的数据接收完成 */
         const FBE = 1 << 2; /* RW 致命总线错误中断 */
@@ -621,17 +586,17 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIDMACStatus {
+impl FlagReg for FsdifDmacStatus {
     const REG: u32 = FSDIF_DMAC_STATUS_OFFSET; // 假设 FSDIF_DMAC_STATUS_OFFSET 是对应的寄存器偏移量
 }
 
-pub fn dmac_status_eb_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIDMACStatus>().bits(), 12, 10)
+pub fn dmac_status_eb_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifDmacStatus>().bits(), 12, 10)
 }
 
 /// FSDIF_DMAC_INT_EN_OFFSET Register
 bitflags! {
-    pub struct MCIDMACIntEn: u32 {
+    pub struct FsdifDmacIntEn: u32 {
         const TI = 1 << 0;  /* RW 发送完成中断使能 */
         const RI = 1 << 1;  /* RW 接收完成中断使能 */
         const FBE = 1 << 2; /* RW 总线错误中断使能 */
@@ -644,13 +609,13 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIDMACIntEn {
+impl FlagReg for FsdifDmacIntEn {
     const REG: u32 = FSDIF_DMAC_INT_EN_OFFSET; // 假设 FSDIF_DMAC_INT_EN_OFFSET 是对应的寄存器偏移量
 }
 
 /// FSDIF_CARD_THRCTL_OFFSET Register
 bitflags! {
-    pub struct MCICardThrctl: u32 {
+    pub struct FsdifCardThrctl: u32 {
         const CARDRD = 1 << 0;   /* RW 读卡threshold使能 */
         const BUSY_CLR = 1 << 1; /* RW busy清中断 */
         const CARDWR = 1 << 2;   /* RO 写卡threshold使能 */
@@ -670,27 +635,38 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCICardThrctl {
+impl FlagReg for FsdifCardThrctl {
     const REG: u32 = FSDIF_CARD_THRCTL_OFFSET; // 假设 FSDIF_CARD_THRCTL_OFFSET 是对应的寄存器偏移量
 }
 
-impl From<MCIFifoDepth> for MCICardThrctl {
-    fn from(value: MCIFifoDepth) -> Self {
-        let value: u32 = value as u32;
-        let value: u32 = 1 << value;
-        MCICardThrctl::from_bits_truncate(value)
+impl From<u32> for FsdifCardThrctl {
+    fn from(val: u32) -> Self {
+        FsdifCardThrctl::from_bits_truncate(val)
+    }
+}
+
+pub enum FsdifFifoDepth {
+    Depth8 = 23,
+    Depth16 = 24,
+    Depth32 = 25,
+    Depth64 = 26,
+    Depth128 = 27,
+}
+
+impl FsdifFifoDepth {
+    pub fn card_thrctl_threshold(self) -> u32 {
+        1 << self as u32
     }
 }
 
 // 读卡 Threshold
-pub fn card_thrctl_threshold(reg:MCIReg,n: u32) -> u32 {
-    reg.read_reg::<MCICardThrctl>().bits() & (1<<n)
+pub fn card_thrctl_threshold(reg:FsdifReg,n: u32) -> u32 {
+    reg.read_reg::<FsdifCardThrctl>().bits() & (1<<n)
 }
 
 /// FSDIF_CLK_SRC_OFFSET Register
 bitflags! {
-    #[derive(Clone, Copy)]
-    pub struct MCIClkSrc: u32 {
+    pub struct FsdifClkSrc: u32 {
         const UHS_EXT_MMC_VOLT = 1 << 0;         /* RW 1.2V供电选择 */
         const UHS_EXT_CLK_ENA = 1 << 1;          /* RW 外部时钟，CIU时钟使能 */
         const UHS_EXT_CLK_MUX = 1 << 31;         /* RW 外部时钟选择 */
@@ -721,11 +697,11 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIClkSrc {
+impl FlagReg for FsdifClkSrc {
     const REG: u32 = FSDIF_CLK_SRC_OFFSET; // 假设 FSDIF_CLK_SRC_OFFSET 是对应的寄存器偏移量
 }
 
-impl MCIClkSrc {
+impl FsdifClkSrc {
     pub fn uhs_clk_div(x: u32) -> Self {
         Self::UHS_CLK_DIV_MASK & Self::from_bits_truncate(x << 8)
     }
@@ -745,42 +721,42 @@ impl MCIClkSrc {
     }
 }
 
-pub fn uhs_clk_div_set(reg: MCIReg, x: u32) {
-    reg.modify_reg::<MCIClkSrc>(|reg| {
-        reg | MCIClkSrc::from_bits_truncate(set_reg32_bits!(x,14,8))
+pub fn uhs_clk_div_set(reg: FsdifReg, x: u32) {
+    reg.modify_reg::<FsdifClkSrc>(|reg| {
+        reg | FsdifClkSrc::from_bits_truncate(set_reg32_bits!(x,14,8))
     });
 }
 
-pub fn uhs_clk_div_get(reg: MCIReg) -> u32 {
-    get_reg32_bits!(reg.read_reg::<MCIClkSrc>().bits(),14,8)
+pub fn uhs_clk_div_get(reg: FsdifReg) -> u32 {
+    get_reg32_bits!(reg.read_reg::<FsdifClkSrc>().bits(),14,8)
 }
 
-pub fn uhs_clk_samp_set(reg: MCIReg, x: u32) {
-    reg.modify_reg::<MCIClkSrc>(|reg| {
-        reg | MCIClkSrc::from_bits_truncate(set_reg32_bits!(x,22,16))
+pub fn uhs_clk_samp_set(reg: FsdifReg, x: u32) {
+    reg.modify_reg::<FsdifClkSrc>(|reg| {
+        reg | FsdifClkSrc::from_bits_truncate(set_reg32_bits!(x,22,16))
     });
 }
 
-pub fn uhs_clk_drv_set(reg: MCIReg, x: u32) {
-    reg.modify_reg::<MCIClkSrc>(|reg| {
-        reg | MCIClkSrc::from_bits_truncate(set_reg32_bits!(x,30,24))
+pub fn uhs_clk_drv_set(reg: FsdifReg, x: u32) {
+    reg.modify_reg::<FsdifClkSrc>(|reg| {
+        reg | FsdifClkSrc::from_bits_truncate(set_reg32_bits!(x,30,24))
     });
 }
 
 /// FSDIF_EMMC_DDR_REG_OFFSET Register
 bitflags! {
-    pub struct MCIEmmcDdrReg: u32 {
+    pub struct FsdifEmmcDdrReg: u32 {
         const CYCLE = 1 << 0; /* RW 1: start bit小于一个周期，0：start bit 为一个周期 */
     }
 }
 
-impl FlagReg for MCIEmmcDdrReg {
+impl FlagReg for FsdifEmmcDdrReg {
     const REG: u32 = FSDIF_EMMC_DDR_REG_OFFSET; // 假设 FSDIF_EMMC_DDR_REG_OFFSET 是对应的寄存器偏移量
 }
 
 /// FSDIF_DESC_LIST_ADDRH_OFFSET Register
 bitflags! {
-    pub struct MCIDescListAddrH: u32 {
+    pub struct FsdifDescListAddrH: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -816,13 +792,13 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIDescListAddrH {
+impl FlagReg for FsdifDescListAddrH {
     const REG: u32 = FSDIF_DESC_LIST_ADDRH_OFFSET; // 假设 FSDIF_DESC_LIST_ADDRH_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_DESC_LIST_ADDRL_OFFSET Register
+/// FSDIF_DESC_LIST_ADDRL_OFFSET Register
 bitflags! {
-    pub struct MCIDescListAddrL: u32 {
+    pub struct FsdifDescListAddrL: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -858,7 +834,7 @@ bitflags! {
     }
 }
 
-impl FlagReg for MCIDescListAddrL {
+impl FlagReg for FsdifDescListAddrL {
     const REG: u32 = FSDIF_DESC_LIST_ADDRL_OFFSET; // 假设 FSDIF_DESC_LIST_ADDRL_OFFSET 是对应的寄存器偏移量
 }
 
@@ -905,7 +881,7 @@ impl FlagReg for FsdifData {
 
 /// FSDIF_BYT_CNT_OFFSET Register
 bitflags! {
-    pub struct MCIBytCnt: u32 {
+    pub struct FsdifBytCnt: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -940,13 +916,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIBytCnt {
+impl FlagReg for FsdifBytCnt {
     const REG: u32 = FSDIF_BYT_CNT_OFFSET; // 假设 FSDIF_BYT_CNT_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_BLK_SIZ_OFFSET Register
+/// FSDIF_BLK_SIZ_OFFSET Register
 bitflags! {
-    pub struct MCIBlkSiz: u32 {
+    pub struct FsdifBlkSiz: u32 {
         const BIT0 = 1 << 0; /* RW 1: 512字节块大小，0：512字节块大小 */
         const BIT1 = 1 << 1; /* RW 1: 512字节块大小，0：512字节块大小 */
         const BIT2 = 1 << 2; /* RW 1: 512字节块大小，0：512字节块大小 */
@@ -982,13 +958,13 @@ bitflags! {
         const ALL_BITS = 0xFFFFFFFF;
     }
 }
-impl FlagReg for MCIBlkSiz {
+impl FlagReg for FsdifBlkSiz {
     const REG: u32 = FSDIF_BLK_SIZ_OFFSET; // 假设 FSDIF_BLK_SIZ_OFFSET 是对应的寄存器偏移量
 }
 
-// FSDIF_TRAN_CARD_CNT_OFFSET Register
+/// FSDIF_TRAN_CARD_CNT_OFFSET Register
 bitflags! {
-    pub struct MCITranCardCnt:u32 {
+    pub struct FsdifTranCardCnt:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1023,13 +999,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCITranCardCnt {
+impl FlagReg for FsdifTranCardCnt {
     const REG: u32 = FSDIF_TRAN_CARD_CNT_OFFSET;
 }
 
 /// FSDIF_TRAN_FIFO_CNT_OFFSET Register
 bitflags! {
-    pub struct MCITranFifoCnt:u32 {
+    pub struct FsdifTranFifoCnt:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1064,13 +1040,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCITranFifoCnt {
+impl FlagReg for FsdifTranFifoCnt {
     const REG: u32 = FSDIF_TRAN_FIFO_CNT_OFFSET;
 }
 
 /// FSDIF_RESP0_OFFSET Register
 bitflags! {
-    pub struct MCIResp0:u32 {
+    pub struct FsdifResp0:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1105,13 +1081,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIResp0 {
+impl FlagReg for FsdifResp0 {
     const REG: u32 = FSDIF_RESP0_OFFSET;
 }
 
 /// FSDIF_RESP1_OFFSET Register
 bitflags! {
-    pub struct MCIResp1:u32 {
+    pub struct FsdifResp1:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1146,13 +1122,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIResp1 {
+impl FlagReg for FsdifResp1 {
     const REG: u32 = FSDIF_RESP1_OFFSET;
 }
 
 /// FSDIF_RESP2_OFFSET Register
 bitflags! {
-    pub struct MCIResp2:u32 {
+    pub struct FsdifResp2:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1187,13 +1163,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIResp2 {
+impl FlagReg for FsdifResp2 {
     const REG: u32 = FSDIF_RESP2_OFFSET;
 }
 
 /// FSDIF_RESP3_OFFSET Register
 bitflags! {
-    pub struct MCIResp3:u32 {
+    pub struct FsdifResp3:u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1228,13 +1204,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIResp3 {
+impl FlagReg for FsdifResp3 {
     const REG: u32 = FSDIF_RESP3_OFFSET;
 }
 
-// FSDIF_CMD_ARG_OFFSET Register
+/// FSDIF_CMD_ARG_OFFSET Register
 bitflags! {
-    pub struct MCICmdArg: u32 {
+    pub struct FsdifCmdArg: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1269,14 +1245,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-
-impl FlagReg for MCICmdArg {
+impl FlagReg for FsdifCmdArg {
     const REG: u32 = FSDIF_CMD_ARG_OFFSET;
 }
 
-// FSDIF_DEBNCE_OFFSET Register
+/// FSDIF_DEBNCE_OFFSET Register
 bitflags! {
-    pub struct MCIDebnce: u32 {
+    pub struct FsdifDebnce: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1311,13 +1286,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIDebnce {
+impl FlagReg for FsdifDebnce {
     const REG: u32 = FSDIF_DEBNCE_OFFSET;
 }
 
-// FSDIF_UID_OFFSET Register
+/// FSDIF_UID_OFFSET Register
 bitflags! {
-    pub struct MCIUid: u32 {
+    pub struct FsdifUid: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1352,13 +1327,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIUid {
+impl FlagReg for FsdifUid {
     const REG: u32 = FSDIF_UID_OFFSET;
 }
 
-// FSDIF_VID_OFFSET Register
+/// FSDIF_VID_OFFSET Register
 bitflags! {
-    pub struct MCIVid: u32 {
+    pub struct FsdifVid: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1393,13 +1368,13 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIVid {
+impl FlagReg for FsdifVid {
     const REG: u32 = FSDIF_VID_OFFSET;
 }
 
 /// FSDIF_HWCONF_OFFSET Register
 bitflags! {
-    pub struct MCIHwconf: u32 {
+    pub struct FsdifHwconf: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1434,12 +1409,12 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIHwconf {
+impl FlagReg for FsdifHwconf {
     const REG: u32 = FSDIF_HWCONF_OFFSET;
 }
 /// FSDIF_CUR_DESC_ADDRL_OFFSET Register
 bitflags! {
-    pub struct MCICurDescAddrL: u32 {
+    pub struct FsdifCurDescAddrL: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1474,12 +1449,12 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCICurDescAddrL {
+impl FlagReg for FsdifCurDescAddrL {
     const REG: u32 = FSDIF_CUR_DESC_ADDRL_OFFSET;
 }
 /// FSDIF_CUR_DESC_ADDRH_OFFSET Register
 bitflags! {
-    pub struct MCIDescAddrH: u32 {
+    pub struct FsdifCurDescAddrH: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1514,12 +1489,12 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIDescAddrH {
+impl FlagReg for FsdifCurDescAddrH {
     const REG: u32 = FSDIF_CUR_DESC_ADDRH_OFFSET;
 }
 /// FSDIF_CUR_BUF_ADDRL_OFFSET Register
 bitflags! {
-    pub struct MCICurBufAddrL: u32 {
+    pub struct FsdifCurBufAddrL: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1554,12 +1529,12 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCICurBufAddrL {
+impl FlagReg for FsdifCurBufAddrL {
     const REG: u32 = FSDIF_CUR_BUF_ADDRL_OFFSET;
 }
 /// FSDIF_CUR_BUF_ADDRH_OFFSET Register
 bitflags! {
-    pub struct MCIBufAddrH: u32 {
+    pub struct FsdifCurBufAddrH: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1594,12 +1569,12 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIBufAddrH {
+impl FlagReg for FsdifCurBufAddrH {
     const REG: u32 = FSDIF_CUR_BUF_ADDRH_OFFSET;
 }
 /// FSDIF_ENABLE_SHIFT_OFFSET Register
 bitflags! {
-    pub struct MCIEnableShift: u32 {
+    pub struct FsdifEnableShift: u32 {
         const BIT0 = 1 << 0;
         const BIT1 = 1 << 1;
         const BIT2 = 1 << 2;
@@ -1634,6 +1609,6 @@ bitflags! {
         const BIT31 = 1 << 31;
     }
 }
-impl FlagReg for MCIEnableShift {
+impl FlagReg for FsdifEnableShift {
     const REG: u32 = FSDIF_ENABLE_SHIFT_OFFSET;
 }
